final/backend/requirements.txt
Django==1.10.1
django-cors-middleware==1.3.1
djangorestframework==3.4.6
mysqlclient==1.3.7
requests==2.11.1
